requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pika
python-dotenv
python-dateutil
pydantic
pydantic-settings